requirements.txt
# Основные зависимости
requests>=2.31.0
python-dotenv>=1.0.0

# CLI интерфейс
click>=8.1.7
colorama>=0.4.6

# Валидация данных (опционально)
pydantic>=2.5.0

# Тестирование
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.24.0

# Разработка
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
